+++ client/game/common/Spinner/Spinner.spec.js
@@ -15,6 +15,10 @@ describe("Spinner.vue", () => {
         expect(cmp.isVueInstance()).toBeTruthy();
     });
 
+    it("has the expected html structure", () => {
+        expect(cmp.vm.$el).toMatchSnapshot();
+    });
+
     it("uses depth prop as border-width", () => {
         cmp.setProps({depth: 10});
         expect(cmp.find(".spinner").attributes().style).toMatch(/border-width: 10px/);